# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Web Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agent/command:
    post:
      operationId: processCommand
      tags: [agent]
      summary: Process a Hinglish/English command
      description: Parse, normalize, and execute an AI agent command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandRequest"
      responses:
        "200":
          description: Command processed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /agent/preview/{filename}:
    get:
      operationId: previewOutput
      tags: [agent]
      summary: Get HTML content of a generated output file
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agent/download/{zipFilename}:
    get:
      operationId: downloadProject
      tags: [agent]
      summary: Download a generated project as a ZIP file
      parameters:
        - name: zipFilename
          in: path
          required: true
          schema:
            type: string
          description: ZIP filename (e.g. login-project-1234567890.zip)
      responses:
        "200":
          description: ZIP file stream
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /agent/session/{sessionId}:
    get:
      operationId: getSession
      tags: [agent]
      summary: Get conversation session state
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationSession"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agent/session/{sessionId}/reset:
    post:
      operationId: resetSession
      tags: [agent]
      summary: Reset a conversation session to start a new task
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationSession"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /agent/versions/{sessionId}/{projectBase}:
    get:
      operationId: getVersionHistory
      tags: [agent]
      summary: Get full version history for a project type within a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
        - name: projectBase
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Version history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionHistory"
        "404":
          description: No versions found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    CommandRequest:
      type: object
      properties:
        command:
          type: string
          description: Natural language command in Hinglish or English
        researchMode:
          type: boolean
          description: Whether to enable internet research mode
          default: false
        sessionId:
          type: string
          nullable: true
          description: Session ID for conversation continuity
      required:
        - command
    ConversationTurn:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, agent]
        content:
          type: string
        timestamp:
          type: string
        outputFile:
          type: string
          nullable: true
        jobId:
          type: string
          nullable: true
        validationScore:
          type: integer
          nullable: true
        confidenceScore:
          type: integer
          nullable: true
        regenerated:
          type: boolean
        followUpQuestions:
          type: array
          items:
            type: string
      required:
        - id
        - role
        - content
        - timestamp
    CollectedConditions:
      type: object
      properties:
        pageType:
          type: string
          nullable: true
        colorScheme:
          type: string
          enum: [light, dark, auto]
          nullable: true
        uiStyle:
          type: array
          items:
            type: string
        hasValidation:
          type: boolean
          nullable: true
        hasNavbar:
          type: boolean
          nullable: true
        hasFooter:
          type: boolean
          nullable: true
        customNotes:
          type: array
          items:
            type: string
        mentionedFeatures:
          type: array
          items:
            type: string
        lastUrl:
          type: string
          nullable: true
    ConversationSession:
      type: object
      properties:
        sessionId:
          type: string
        startedAt:
          type: string
        lastActivityAt:
          type: string
        currentTask:
          type: string
          nullable: true
        taskState:
          type: string
          enum: [idle, collecting, generating, awaiting_feedback, refining, complete]
        collectedConditions:
          $ref: "#/components/schemas/CollectedConditions"
        turnCount:
          type: integer
        generatedOutputs:
          type: array
          items:
            type: string
        followUpQuestions:
          type: array
          items:
            type: string
        recentTurns:
          type: array
          items:
            $ref: "#/components/schemas/ConversationTurn"
      required:
        - sessionId
        - startedAt
        - lastActivityAt
        - taskState
        - collectedConditions
        - turnCount
        - generatedOutputs
        - followUpQuestions
        - recentTurns
    ValidationCheck:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
          enum: [structure, css, js, accessibility, responsiveness, design-system]
        name:
          type: string
        status:
          type: string
          enum: [pass, fail, warn]
        detail:
          type: string
        weight:
          type: integer
      required:
        - id
        - category
        - name
        - status
        - detail
        - weight
    ValidationReport:
      type: object
      properties:
        passed:
          type: boolean
        score:
          type: integer
          description: 0–100 quality score
        grade:
          type: string
          enum: [A, B, C, D, F]
        checks:
          type: array
          items:
            $ref: "#/components/schemas/ValidationCheck"
        errors:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
        canAutoFix:
          type: boolean
      required:
        - passed
        - score
        - grade
        - checks
        - errors
        - warnings
        - suggestions
        - canAutoFix
    ConfidenceDimension:
      type: object
      properties:
        name:
          type: string
        score:
          type: integer
        weight:
          type: integer
        rationale:
          type: string
      required:
        - name
        - score
        - weight
        - rationale
    ConfidenceResult:
      type: object
      properties:
        overall:
          type: integer
          description: 0–100 weighted confidence score
        grade:
          type: string
          enum: [A, B, C, D, F]
        passesThreshold:
          type: boolean
        threshold:
          type: integer
        dimensions:
          type: array
          items:
            $ref: "#/components/schemas/ConfidenceDimension"
        recommendation:
          type: string
          enum: [use, regenerate, fallback]
        regenerationHint:
          type: string
          nullable: true
      required:
        - overall
        - grade
        - passesThreshold
        - threshold
        - dimensions
        - recommendation
    CommandResult:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [success, error, processing]
        normalizedCommand:
          type: string
          description: The normalized English interpretation of the command
        detectedIntent:
          type: string
          description: Detected intent (create, clone, research, analyze, etc.)
        detectedMode:
          type: string
          enum: [analyze, research, direct]
        detectedUrl:
          type: string
          nullable: true
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskStep"
        outputFile:
          type: string
          nullable: true
          description: Generated output filename
        projectSlug:
          type: string
          nullable: true
          description: Packaged project slug (used for ZIP download)
        downloadUrl:
          type: string
          nullable: true
          description: URL to download the full runnable project as a ZIP
        message:
          type: string
        validationReport:
          nullable: true
          $ref: "#/components/schemas/ValidationReport"
        confidenceResult:
          nullable: true
          $ref: "#/components/schemas/ConfidenceResult"
        regenerated:
          type: boolean
          description: Whether the output was regenerated after a low confidence score
        sessionId:
          type: string
          description: Session ID for conversation continuity
        conversationIntent:
          type: string
          enum: [continue, new_task, explicit_reset, follow_up]
          description: Detected intent of this turn relative to the session
        agentMessage:
          type: string
          description: Human-readable agent response for the chat UI
        followUpQuestions:
          type: array
          items:
            type: string
          description: Suggested follow-up questions for the user
        session:
          $ref: "#/components/schemas/ConversationSession"
        responseMode:
          type: string
          enum: [generate, clarify, acknowledge]
          description: "generate = full output produced, clarify = asking user a question, acknowledge = confirming before generating"
        clarificationQuestion:
          type: string
          nullable: true
          description: The single clarification question to ask the user
        quickReplies:
          type: array
          items:
            type: string
          description: Tap-to-send quick reply chip suggestions
        acknowledgment:
          type: string
          nullable: true
          description: Natural acknowledgment opener shown before generating
        webResearchUsed:
          type: boolean
          description: Whether live web sources were fetched for this response
        webSources:
          type: array
          items:
            $ref: "#/components/schemas/WebSourceResult"
          description: Web sources consulted during research
        webCrossCheckedFacts:
          type: array
          items:
            type: string
          description: Facts confirmed by 2+ independent web sources
        webAdaptedPatterns:
          type: array
          items:
            type: string
          description: Project-aligned patterns adapted from web sources (never raw copy-paste)
        chatOperation:
          type: string
          nullable: true
          description: "Chat management operation performed: reset_chat | delete_last | delete_message | new_session"
        userTurnId:
          type: string
          nullable: true
          description: Server-side session turn ID for the user message in this exchange
        agentTurnId:
          type: string
          nullable: true
          description: Server-side session turn ID for the agent message in this exchange
        securityReport:
          nullable: true
          $ref: "#/components/schemas/SecurityReport"
        performanceReport:
          nullable: true
          $ref: "#/components/schemas/PerformanceReport"
        versionId:
          type: string
          nullable: true
          description: "Version ID of this generation (e.g. v1, v2, v3)"
        versionNumber:
          type: integer
          nullable: true
          description: Monotonically increasing version number for this project type in this session
        editApplied:
          type: boolean
          description: Whether this was a targeted edit applied to an existing output
      required:
        - jobId
        - status
        - normalizedCommand
        - detectedIntent
        - detectedMode
        - tasks
        - message
        - regenerated
        - sessionId
        - conversationIntent
        - agentMessage
        - followUpQuestions
        - session
        - responseMode
        - quickReplies
    TaskStep:
      type: object
      properties:
        step:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [pending, running, done, error, skipped]
        detail:
          type: string
          nullable: true
      required:
        - step
        - name
        - status
...
    FileContent:
      type: object
      properties:
        filename:
          type: string
        content:
          type: string
        pageType:
          type: string
      required:
        - filename
        - content
        - pageType
...
    WebSourceResult:
      type: object
      properties:
        url:
          type: string
        domain:
          type: string
        title:
          type: string
        relevanceScore:
          type: number
        qualityScore:
          type: number
        headings:
          type: array
          items:
            type: string
        fetchedAt:
          type: string
      required:
        - url
        - domain
        - title
        - relevanceScore
        - qualityScore
...
    SecurityIssue:
      type: object
      properties:
        id:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        type:
          type: string
        description:
          type: string
        pattern:
          type: string
        suggestion:
          type: string
        autoFixed:
          type: boolean
      required: [id, severity, type, description, pattern, suggestion, autoFixed]
    SecurityReport:
      type: object
      properties:
        safe:
          type: boolean
        score:
          type: integer
          description: 0–100 (100 = no issues)
        issues:
          type: array
          items:
            $ref: "#/components/schemas/SecurityIssue"
        blockedCount:
          type: integer
        summary:
          type: string
      required: [safe, score, issues, blockedCount, summary]
    PerformanceOptimization:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
          enum: [css, html, js, assets, hints]
        description:
          type: string
        bytesSaved:
          type: integer
      required: [id, category, description, bytesSaved]
    PerformanceReport:
      type: object
      properties:
        score:
          type: integer
          description: 0–100
        grade:
          type: string
          enum: [A, B, C, D, F]
        originalSizeBytes:
          type: integer
        optimizedSizeBytes:
          type: integer
        reductionBytes:
          type: integer
        reductionPercent:
          type: integer
        optimizationsApplied:
          type: array
          items:
            $ref: "#/components/schemas/PerformanceOptimization"
        summary:
          type: string
      required: [score, grade, originalSizeBytes, optimizedSizeBytes, reductionBytes, reductionPercent, optimizationsApplied, summary]
    ProjectVersionSummary:
      type: object
      properties:
        versionId:
          type: string
        versionNumber:
          type: integer
        command:
          type: string
        changesSummary:
          type: string
        linesAdded:
          type: integer
        linesRemoved:
          type: integer
        createdAt:
          type: string
        scores:
          type: object
          properties:
            validation:
              type: integer
            confidence:
              type: integer
            security:
              type: integer
            performance:
              type: integer
      required: [versionId, versionNumber, command, changesSummary, linesAdded, linesRemoved, createdAt]
    VersionHistory:
      type: object
      properties:
        projectBaseName:
          type: string
        sessionId:
          type: string
        currentVersion:
          type: integer
        currentVersionId:
          type: string
        totalVersions:
          type: integer
        versions:
          type: array
          items:
            $ref: "#/components/schemas/ProjectVersionSummary"
        lastUpdated:
          type: string
      required: [projectBaseName, sessionId, currentVersion, currentVersionId, totalVersions, versions, lastUpdated]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
          nullable: true
      required:
        - error